requirements.txt
fastapi==0.111.2
uvicorn==0.24.0
pyserial==3.6
